Added a recommend service
prepareRecommend: checks if the request is formatted properly, with the intention to delegate this to the DB once done (TODO). It currently returns a stub recommendation object.

notifyRecommend: Still work in progress, we need to clarify the api specification from the front end, as a post request apparently does not allow the specified format.