Added ML requested features
Every new portfolio will have an int array, which is empty by default.

New actions that interact with it (all of them through post to user-portfolio service):

resetPortfolioReommend - replace the recommend field with a new int array. No additional data required.

pushPortfolioRecommend - push a value to the recommend array. It is tested to be a number. Requires 'value' field in the request body with a numeric value.

setPortfolioRecommend - replace the array with a given one. Requires 'values' field in the request body, with an array of int. The elements are NOT tested to be int.